Segment pack/unpack
