+++ src/mpid/ch3/include/mpid_rma_shm.h
@@ -380,7 +380,7 @@ static inline int MPIDI_CH3I_Shm_acc_op(const void *origin_addr, int origin_coun
         MPIR_ERR_CHKANDJUMP1(seg == NULL, mpi_errno, MPI_ERR_OTHER, "**nomem", "**nomem %s",
                              "MPIDU_Segment");
         MPIR_Segment_init(origin_addr, origin_count, origin_datatype, seg, 0);
-        MPIDU_Segment_pack(seg, first, &last, packed_buf);
+        MPIR_Segment_pack(seg, first, &last, packed_buf);
         MPIR_Segment_free(seg);
 
         MPIR_Datatype_is_contig(basic_type, &is_predef_contig);
@@ -527,7 +527,7 @@ static inline int MPIDI_CH3I_Shm_get_acc_op(const void *origin_addr, int origin_
         MPIR_ERR_CHKANDJUMP1(seg == NULL, mpi_errno, MPI_ERR_OTHER, "**nomem", "**nomem %s",
                              "MPIDU_Segment");
         MPIR_Segment_init(origin_addr, origin_count, origin_datatype, seg, 0);
-        MPIDU_Segment_pack(seg, first, &last, packed_buf);
+        MPIR_Segment_pack(seg, first, &last, packed_buf);
         MPIR_Segment_free(seg);
 
         MPIR_Datatype_is_contig(basic_type, &is_predef_contig);
