+++ src/mpid/ch3/src/ch3u_buffer.c
@@ -87,7 +87,7 @@ void MPIDI_CH3U_Buffer_copy(
 	last = sdata_sz;
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
                           "pre-unpack last=%" PRIdPTR, last ));
-	MPIDU_Segment_unpack(&seg, 0, &last, (char*)sbuf + sdt_true_lb);
+	MPIR_Segment_unpack(&seg, 0, &last, (char*)sbuf + sdt_true_lb);
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
 			 "pre-unpack last=%" PRIdPTR, last ));
 	/* --BEGIN ERROR HANDLING-- */
@@ -108,7 +108,7 @@ void MPIDI_CH3U_Buffer_copy(
 	last = sdata_sz;
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
 			       "pre-pack last=%" PRIdPTR, last ));
-	MPIDU_Segment_pack(&seg, 0, &last, (char*)rbuf + rdt_true_lb);
+	MPIR_Segment_pack(&seg, 0, &last, (char*)rbuf + rdt_true_lb);
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
 			    "post-pack last=%" PRIdPTR, last ));
 	/* --BEGIN ERROR HANDLING-- */
@@ -165,7 +165,7 @@ void MPIDI_CH3U_Buffer_copy(
 	    MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
                "pre-pack first=%" PRIdPTR ", last=%" PRIdPTR,
 						sfirst, last ));
-	    MPIDU_Segment_pack(&sseg, sfirst, &last, buf + buf_off);
+	    MPIR_Segment_pack(&sseg, sfirst, &last, buf + buf_off);
 	    MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
                "post-pack first=%" PRIdPTR ", last=%" PRIdPTR,
                sfirst, last ));
@@ -179,7 +179,7 @@ void MPIDI_CH3U_Buffer_copy(
 	    MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
              "pre-unpack first=%" PRIdPTR ", last=%" PRIdPTR,
 						rfirst, last ));
-	    MPIDU_Segment_unpack(&rseg, rfirst, &last, buf);
+	    MPIR_Segment_unpack(&rseg, rfirst, &last, buf);
 	    MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_OTHER,VERBOSE,(MPL_DBG_FDEST,
              "post-unpack first=%" PRIdPTR ", last=%" PRIdPTR,
 						rfirst, last ));
