+++ src/mpid/ch3/src/ch3u_request.c
@@ -99,7 +99,7 @@ int MPIDI_CH3U_Request_load_send_iov(MPIR_Request * const sreq,
     MPIR_Assert(sreq->dev.segment_first < last);
     MPIR_Assert(last > 0);
     MPIR_Assert(*iov_n > 0 && *iov_n <= MPL_IOV_LIMIT);
-    MPIDU_Segment_pack_vector(sreq->dev.segment_ptr, sreq->dev.segment_first, 
+    MPIR_Segment_pack_vector(sreq->dev.segment_ptr, sreq->dev.segment_first, 
 			     &last, iov, iov_n);
     MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_CHANNEL,VERBOSE,(MPL_DBG_FDEST,
     "post-pv: first=%" PRIdPTR ", last=%" PRIdPTR ", iov_n=%d",
@@ -155,7 +155,7 @@ int MPIDI_CH3U_Request_load_send_iov(MPIR_Request * const sreq,
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_CHANNEL,VERBOSE,(MPL_DBG_FDEST,
                "pre-pack: first=%" PRIdPTR ", last=%" PRIdPTR,
 			  sreq->dev.segment_first, last));
-	MPIDU_Segment_pack(sreq->dev.segment_ptr, sreq->dev.segment_first, 
+	MPIR_Segment_pack(sreq->dev.segment_ptr, sreq->dev.segment_first, 
 			  &last, (char*) sreq->dev.tmpbuf + iov_data_copied);
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_CHANNEL,VERBOSE,(MPL_DBG_FDEST,
               "post-pack: first=%" PRIdPTR ", last=%" PRIdPTR,
@@ -267,7 +267,7 @@ int MPIDI_CH3U_Request_load_recv_iov(MPIR_Request * const rreq)
 			  rreq->dev.segment_first, last, rreq->dev.iov_count));
 	MPIR_Assert(rreq->dev.segment_first < last);
 	MPIR_Assert(last > 0);
-	MPIDU_Segment_unpack_vector(rreq->dev.segment_ptr, 
+	MPIR_Segment_unpack_vector(rreq->dev.segment_ptr, 
 				   rreq->dev.segment_first,
 				   &last, &rreq->dev.iov[0], &rreq->dev.iov_count);
 	MPL_DBG_MSG_FMT(MPIDI_CH3_DBG_CHANNEL,VERBOSE,(MPL_DBG_FDEST,
@@ -419,7 +419,7 @@ int MPIDI_CH3U_Request_unpack_srbuf(MPIR_Request * rreq)
 	tmpbuf_last = (int)rreq->dev.segment_size;
     }
     last = tmpbuf_last;
-    MPIDU_Segment_unpack(rreq->dev.segment_ptr, rreq->dev.segment_first, 
+    MPIR_Segment_unpack(rreq->dev.segment_ptr, rreq->dev.segment_first, 
 			&last, rreq->dev.tmpbuf);
     if (last == 0 || last == rreq->dev.segment_first)
     {
@@ -538,7 +538,7 @@ int MPIDI_CH3U_Request_unpack_uebuf(MPIR_Request * rreq)
 	    MPIR_Segment_init(rreq->dev.user_buf, rreq->dev.user_count, 
 			      rreq->dev.datatype, &seg, 0);
 	    last = unpack_sz;
-	    MPIDU_Segment_unpack(&seg, 0, &last, rreq->dev.tmpbuf);
+	    MPIR_Segment_unpack(&seg, 0, &last, rreq->dev.tmpbuf);
 	    if (last != unpack_sz)
 	    {
 		/* --BEGIN ERROR HANDLING-- */
