+++ src/mpid/pamid/src/coll/alltoall/mpido_alltoall.c
@@ -347,7 +347,7 @@ int MPIDO_Alltoall_simple(const void *sendbuf,
       }
       DLOOP_Offset last = send_size*size;
       MPIR_Segment_init(sendbuf, sendcount*size, sendtype, &segment, 0);
-      MPIDU_Segment_pack(&segment, 0, &last, snd_noncontig_buff);
+      MPIR_Segment_pack(&segment, 0, &last, snd_noncontig_buff);
 
     }
   }
