+++ src/mpid/pamid/src/coll/gather/mpido_gather.c
@@ -495,7 +495,7 @@ int MPIDO_Gather_simple(const void *sendbuf,
         }
         DLOOP_Offset last = send_size;
         MPIR_Segment_init(sendbuf, sendcount, sendtype, &segment, 0);
-        MPIDU_Segment_pack(&segment, 0, &last, snd_noncontig_buff);
+        MPIR_Segment_pack(&segment, 0, &last, snd_noncontig_buff);
     }
   }
   else
