+++ src/mpid/pamid/src/coll/gatherv/mpido_gatherv.c
@@ -379,7 +379,7 @@ int MPIDO_Gatherv_simple(const void *sendbuf,
       }
       DLOOP_Offset last = send_size;
       MPIR_Segment_init(sendbuf, sendcount, sendtype, &segment, 0);
-      MPIDU_Segment_pack(&segment, 0, &last, snd_noncontig_buff);
+      MPIR_Segment_pack(&segment, 0, &last, snd_noncontig_buff);
     }
   }
   else
